fix(zentao): update progress type (#3767)
* fix(jenkins): update build stages logic (#3760)

* feat(dora): update logic for change lead time (#3742)

* feat(dora): update logic for change lead time

closes #3516

* feat(dora): update e2e according to review

* fix(zentao): update progress typoe

+++ plugins/zentao/models/archived/execution.go
@@ -80,8 +80,8 @@ type ZentaoExecution struct {
 	TotalConsumed  int    `json:"totalConsumed"`
 	TotalLeft      int    `json:"totalLeft"`
 	ProjectId      uint64
-	Progress       int  `json:"progress"`
-	CaseReview     bool `json:"caseReview"`
+	Progress       float64 `json:"progress"`
+	CaseReview     bool    `json:"caseReview"`
 	archived.NoPKModel
 }
